requirements.txt
Flask
qrcode
Flask-Mail
python-dotenv
Flask-WTF
Flask-Login
Flask-CORS
Pillow
gunicorn
requests